+++ Packages/Tracking/Examples~/Example Assets/3D Interaction/Pinch To Paint/Scripts/PaintCursor.cs
@@ -6,11 +6,9 @@
  * between Ultraleap and you, your company or other organization.             *
  ******************************************************************************/
 
-using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 
-namespace Leap.Unity.Preview
+namespace Leap.Unity.Examples
 {
     /// <summary>
     /// The paint cursor deals with the visuals of the paint cursor,