+++ Packages/Tracking/Examples~/Example Assets/3D Interaction/Pinch To Paint/Scripts/ThickRibbonRenderer.cs
@@ -6,12 +6,11 @@
  * between Ultraleap and you, your company or other organization.             *
  ******************************************************************************/
 
-using System;
 using System.Collections.Generic;
 using UnityEngine;
 
 
-namespace Leap.Unity.Preview
+namespace Leap.Unity.Examples
 {
 
     /// <summary>