+++ Packages/Tracking/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [NEXT] - unreleased
 
 ### Added
+- Paint To Pinch example scene
+- Explanation text to all XR example scenes
 - Turntable and Pullcord example scene
 
 ### Changed